+++ docs/Whats-New.md
@@ -138,6 +138,7 @@ Vanilla fixes:
 - Fix incorrect merging of sections and keys in INI files (by ZivDero)
 - Fix game end text being stretched horizontally (by ZivDero)
 - Fix a bug where upon restoring focus to the window the last theme would play twice (by ZivDero)
+- Fix a bug where units that had AA-capable secondary weapons but AA-uncapable primary weapons did not automatically fire at aircraft (by Rampastring)
 
 :::